% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfalls_Wilcoxon.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{arranges single legend}
\usage{
grid_arrange_shared_legend(
  ...,
  title = "title",
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\value{
grid.grab object
}
\description{
arranges single legend
}
