% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_histograms.R
\name{p_value_one_histogram}
\alias{p_value_one_histogram}
\title{plots p-value histogram for a list of p-values
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion}
\usage{
p_value_one_histogram(p_values, name, b_param = 0.01)
}
\arguments{
\item{p_values}{p-values vector}

\item{name}{name of the data set used to produce p-values}

\item{b_param}{the threshold parameter - the probablity that the number of p-values in each bin will exceed the plotted level, default is 1\%}
}
\value{
p-values histogram
}
\description{
plots p-value histogram for a list of p-values
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion
}
