% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{plot enrichment analysis for one potential SL gene pairs
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion}
\usage{
plot_enrichment(sl_pair, sens_matrix1, alt_matrix1, name)
}
\arguments{
\item{sl_pair}{geneA and geneB vector}

\item{name}{name of the knock-down experiment}

\item{sens_matrix}{geneB dependency matrix in CCLE cell lines for a given knock-down experiment}

\item{alt_matrix}{geneA alteration matrix for CCLE cell lines}
}
\value{
enrichment plot
}
\description{
plot enrichment analysis for one potential SL gene pairs
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion
}
