% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_histograms.R
\name{plot_many_histograms}
\alias{plot_many_histograms}
\title{plots 4 p-value histograms}
\usage{
plot_many_histograms(
  p_values_list,
  data_set_names,
  analysis_type,
  b_param = 0.01
)
}
\arguments{
\item{p_values_list}{list of p-values vectors from 4 different testing setups}

\item{data_set_names}{names for each testing setup}

\item{b_param}{the threshold parameter - the probablity that the number of p-values in each bin will exceed the plotted level, default is 1\%}
}
\value{
p-values histograms
}
\description{
plots 4 p-value histograms
}
