% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wilcoxon.R
\name{plot_Wilcoxon_rank_sum_test}
\alias{plot_Wilcoxon_rank_sum_test}
\title{plot boxplot for Wilcoxon for potential geneA - geneB synthetic lethal pair
check if geneA altered cell lines are more dependent on geneB}
\usage{
plot_Wilcoxon_rank_sum_test(
  sl_pair,
  sens_matrix1,
  alt_matrix1,
  experiment = "",
  pval = F
)
}
\arguments{
\item{sl_pair}{c(geneA_ENSEMBL, geneB_ENSEMBL) from a list of potential SL gene pairs}

\item{sens_matrix1}{geneB dependency matrix in CCLE cell lines for a given knock-down experiment}

\item{alt_matrix1}{geneA alteration matrix for CCLE cell lines}

\item{experiment}{name of the experiment}

\item{pval}{p-value from the Wilcoxon test}
}
\value{
boxplot of the Wilcoxon test
}
\description{
plot boxplot for Wilcoxon for potential geneA - geneB synthetic lethal pair
check if geneA altered cell lines are more dependent on geneB
}
