% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfalls_Wilcoxon.R
\name{primary_site}
\alias{primary_site}
\title{checks primary site of a cell line
where geneA which is altered in the minimum no of cell lines for Wilcoxon test
and geneB is present in the chosen knock-down experiment}
\usage{
primary_site(cl_id, cl_ann)
}
\arguments{
\item{cl_id}{cell line id}

\item{cl_ann}{cell line annotations}
}
\value{
primary site
}
\description{
checks primary site of a cell line
where geneA which is altered in the minimum no of cell lines for Wilcoxon test
and geneB is present in the chosen knock-down experiment
}
