% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{verify_enrichment}
\alias{verify_enrichment}
\title{enrichment analysis for list of potential SL gene pairs
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion}
\usage{
verify_enrichment(
  sl_pairs,
  sens_matrix = list(SLiCell::sens_Drive_RSA, SLiCell::sens_Drive_Ataris,
    SLiCell::sens_Achilles_AC, SLiCell::sens_Achilles_Demeter),
  data_set_names = c("Drive_RSA", "Drive_Ataris", "Achilles_AC", "Achilles_Demeter"),
  alt_matrix1 = SLiCell::gene_cl_mut_matrix
)
}
\arguments{
\item{sl_pairs}{list of geneA geneB pairs for verifing}

\item{sens_matrix}{list of geneB dependency matrices in geneA altered cell lines, default - all experimental data preloaded in the package}

\item{data_set_names}{names for the dependency data sets - will be visible in results}

\item{alt_matrix1}{geneA alteration matrix for cell lines, default - mutation_matrix preloaded in the package}
}
\value{
list of enrichment scores and their p-values
}
\description{
enrichment analysis for list of potential SL gene pairs
enrichment set - geneA altered cell lines
geneB sensitivity is a ranking criterion
}
