% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wilcoxon.R
\name{verify_sl_pairs}
\alias{verify_sl_pairs}
\title{verify list of potential SL gene pairs
performs Wilcoxon test for each pair
where geneA which is altered in the minimum no of cell lines for Wilcoxon test
and geneB is present in the chosen knock-down experiment}
\usage{
verify_sl_pairs(
  sl_pair_list,
  sens_matrix = list(SLiCell::sens_Drive_RSA, SLiCell::sens_Drive_Ataris,
    SLiCell::sens_Achilles_AC, SLiCell::sens_Achilles_Demeter),
  data_set_names = c("Drive_RSA", "Drive_Ataris", "Achilles_AC", "Achilles_Demeter"),
  alt_matrix1 = SLiCell::gene_cl_mut_matrix,
  min_alt_no = 10
)
}
\arguments{
\item{sl_pair_list}{list of geneA geneB pairs for verifing}

\item{sens_matrix}{list of geneB dependency matrices in geneA altered cell lines, default - all experimental data preloaded in the package}

\item{data_set_names}{names for the dependency data sets - will be visible in results}

\item{alt_matrix1}{geneA alteration matrix for cell lines, default - mutation_matrix preloaded in the package}

\item{min_alt_no}{minimum no of cell lines which have geneA alteration, default 10}
}
\value{
list of Wilcoxon rank-sum test results separately for each dependency data set
}
\description{
verify list of potential SL gene pairs
performs Wilcoxon test for each pair
where geneA which is altered in the minimum no of cell lines for Wilcoxon test
and geneB is present in the chosen knock-down experiment
}
