% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfalls_Wilcoxon.R
\name{wf}
\alias{wf}
\title{creates one waterfall plot
where geneA altered cell lines (for Wilcoxon test subpopulation) are colored
and geneB depenency is plotted}
\usage{
wf(
  sl_pair,
  sens,
  alt,
  experiment,
  pval = F,
  pallete = SLiCell::sites_colors,
  cl_annotations = SLiCell::cl_annotations
)
}
\arguments{
\item{sl_pair}{geneA geneB pair for verifing}

\item{sens}{list of geneB dependency matrices in geneA altered cell lines, default - all experimental data preloaded in the package}

\item{alt}{geneA alteration matrix for cell lines, default - mutation_matrix preloaded in the package}

\item{experiment}{name of the dependency data set}
}
\value{
waterfall plot
}
\description{
creates one waterfall plot
where geneA altered cell lines (for Wilcoxon test subpopulation) are colored
and geneB depenency is plotted
}
